+++ app/src/main/java/com/mohamedsayed/theair/viewmodel/TvShowDetailsVModel.kt
@@ -11,15 +11,15 @@ import com.mohamedsayed.theair.model.network.RetrofitService
 
 class TvShowDetailsVModel(val apiService: ApiService) : ViewModel() {
 
-    fun getTvShowDetails(tv_id: Long): LiveData<ApiResponse<TvShowDetails>> {
+    fun getTvShowDetails(tv_id: Int): LiveData<ApiResponse<TvShowDetails>> {
         return apiService.getTvShowDetails(tv_id)
     }
 
-    fun rateTvShow(tv_id: Long,value:Double): LiveData<ApiResponse<String>> {
+    fun rateTvShow(tv_id: Int,value:Double): LiveData<ApiResponse<String>> {
         return apiService.rateTvShow(tv_id,value)
     }
 
-    fun getTvShowSimilar(tv_id: Long): LiveData<ApiResponse<TvShowResults>> {
+    fun getTvShowSimilar(tv_id: Int): LiveData<ApiResponse<TvShowResults>> {
         return apiService.getTvShowSimilar(tv_id)
     }
 }
